refactor(libsinsp): move transformers in differnt files
